# src/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "B4 network packet processor REST API",
        "title": "B4 API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/devices": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get all devices",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.DevicesResponse"
                        }
                    }
                }
            }
        },
        "/devices/{mac}/alias": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Manage device alias (get, set, or delete)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MAC address",
                        "name": "mac",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/devices/{mac}/vendor": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Devices"
                ],
                "summary": "Get device vendor info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "MAC address",
                        "name": "mac",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handler.VendorInfo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "handler.DeviceInfo": {
            "type": "object",
            "properties": {
                "alias": {
                    "type": "string"
                },
                "hostname": {
                    "type": "string"
                },
                "ip": {
                    "type": "string"
                },
                "is_private": {
                    "type": "boolean"
                },
                "mac": {
                    "type": "string"
                },
                "vendor": {
                    "type": "string"
                }
            }
        },
        "handler.DevicesResponse": {
            "type": "object",
            "properties": {
                "available": {
                    "type": "boolean"
                },
                "devices": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/handler.DeviceInfo"
                    }
                },
                "source": {
                    "type": "string"
                }
            }
        },
...
        "handler.VendorInfo": {
            "type": "object",
            "properties": {
                "company": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter \"Bearer {token}\" to authorize",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}